requirements.txt
Flask==3.0.0
Flask-Limiter==3.5.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
